pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = ""
authors = [
    {name = "mammadov115",email = "mehman.mammadov182@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django == 5.2.4",
    "djangorestframework == 3.16.0",
    "dotenv == 0.9.9",
    "requests == 2.32.4",
    "psycopg2-binary == 2.9.10",
    "pytest (>=8.4.1,<9.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
]

[tool.poetry]
packages = [{include = "project", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.11.1"
pytest-mock = "^3.14.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

